+++ examples/nextjs-swc/src/app/[lang]/app-router-demo/page.tsx
@@ -1,0 +1,8 @@
+import { HomePage } from '../../../components/HomePage'
+import { initLingui } from '../../../initLingui'
+
+export default async function Page(props) {
+  const lang = (await props.params).lang
+  initLingui(lang)
+  return <HomePage />
+}