+++ examples/react-native/src/MainScreen.tsx
@@ -1,13 +1,13 @@
 import React, { useState } from "react";
 import { StyleSheet, Text, View, Alert, SafeAreaView } from "react-native";
-import { Plural, SelectOrdinal, t, Trans } from "@lingui/macro";
-import { useLingui } from "@lingui/react";
+import { Plural, SelectOrdinal, Trans } from "@lingui/react/macro";
+import { useLingui } from "@lingui/react/macro";
 import { Button } from "./PaddedButton";
 import { Heading } from "./Components";
 
 export const Body = React.memo(() => {
   const [messages, setMessages] = useState<string[]>([]);
-  const { i18n } = useLingui();
+  const { i18n, t } = useLingui();
 
   const markAllAsRead = () => {
     // NOTE - here we're implicitly using the i18n instance that does NOT come from the React context, but from @lingui/core
@@ -31,8 +31,8 @@ export const Body = React.memo(() => {
     const newActiveLanguage = activeLanguage === "en" ? "cs" : "en";
     const catalog =
       newActiveLanguage === "en"
-        ? require("./locales/en/messages.js")
-        : require("./locales/cs/messages.js");
+        ? require("./locales/en/messages.po")
+        : require("./locales/cs/messages.po");
     i18n.load(newActiveLanguage, catalog.messages);
     i18n.activate(newActiveLanguage);
   };
@@ -65,6 +65,8 @@ const Inbox = ({
   markAsRead: () => void;
   addMessage: () => void;
 }) => {
+  const { t } = useLingui();
+
   const messagesCount = messages.length;
 
   return (