+++ packages/cli/src/api/help.ts
@@ -1,6 +1,6 @@
 /**
- * Detect where's is the command lingui extract or lingui compile
- * and how is being run (npm, yarn) and construct help
+ * Detect where is the command lingui extract or lingui compile
+ * and how is being run (npm, yarn, pnpm) and construct help
  * for follow-up commands based on that.
  *
  * Example:
@@ -12,6 +12,10 @@
  * ...
  * (use "yarn lingui compile" to compile catalogs for production)
  *
+ * $ pnpm run extract
+ * ...
+ * (use "pnpm run compile" to compile catalogs for production)
+ *
  * $ npm run extract
  * ...
  * (use "npm run compile" to compile catalogs for production)
@@ -22,7 +26,9 @@ export function helpRun(command: string) {
   let findRootPkgJson: Record<string, unknown>
   try {
     findRootPkgJson = require(resolve(join(process.cwd(), "package.json")))
-  } catch (error) {}
+  } catch (error) {
+    // noting
+  }
 
   if (findRootPkgJson?.scripts) {
     const res = Object.entries(findRootPkgJson.scripts).find(([_, value]) =>
@@ -34,10 +40,25 @@ export function helpRun(command: string) {
     }
   }
 
-  const isYarn =
-    process.env.npm_config_user_agent &&
-    process.env.npm_config_user_agent.includes("yarn")
-  const runCommand = isYarn ? "yarn" : "npm run"
+  const runCommand = runCommandFrom(process.env.npm_config_user_agent)
 
   return `${runCommand} ${command}`
+}
+
+function runCommandFrom(userAgent: string | undefined): string {
+  const defaultRunCommand = "npm run"
+
+  if (!userAgent) {
+    return defaultRunCommand
+  }
+
+  if (userAgent.includes("yarn")) {
+    return "yarn"
+  }
+
+  if (userAgent.includes("pnpm")) {
+    return "pnpm run"
+  }
+
+  return defaultRunCommand
 }