+++ packages/react/src/format.test.tsx
@@ -1,7 +1,6 @@
-import * as React from "react"
 import { render } from "@testing-library/react"
+import * as React from "react"
 import { formatElements } from "./format"
-// eslint-disable-next-line import/no-extraneous-dependencies
 import { mockConsole } from "@lingui/jest-mocks"
 
 describe("formatElements", function () {
@@ -31,6 +30,20 @@ describe("formatElements", function () {
     expect(
       html(formatElements("<0>About</0>", { 0: <a href="/about" /> }))
     ).toEqual('<a href="/about">About</a>')
+  })
+
+  it("should preserve newlines", function () {
+    expect(html(formatElements("<0>Inn\ner</0>", { 0: <strong /> }))).toEqual(
+      "<strong>Inn\ner</strong>"
+    )
+
+    expect(
+      html(formatElements("Before <0>Inn\r\ner</0> After", { 0: <strong /> }))
+    ).toEqual("Before <strong>Inn\r\ner</strong> After")
+
+    expect(
+      html(formatElements("<0>Ab\rout</0>", { 0: <a href="/about" /> }))
+    ).toEqual('<a href="/about">Ab\rout</a>')
   })
 
   it("should preserve named element props", function () {
@@ -70,7 +83,7 @@ describe("formatElements", function () {
         )
       )
     ).toEqual(
-      'Before <a href="/about">Inside <strong>Nested</strong> Between <br> After</a>'
+      'Before \n<a href="/about">Inside <strong>\nNested</strong>\n Between <br> After</a>'
     )
   })
 
@@ -138,13 +151,17 @@ describe("formatElements", function () {
     const elements = formatElements("<div><0/><0/></div>", {
       div: <div />,
       0: <span>hi</span>,
-    }) as Array<React.ReactElement>
+    })
 
-    expect(elements).toHaveLength(1)
-    const childElements = elements[0]!.props.children
-    const childKeys = childElements
-      .map((el: React.ReactElement) => el?.key)
-      .filter(Boolean)
-    expect(cleanPrefix(childKeys[0])).toBeLessThan(cleanPrefix(childKeys[1]))
+    expect(React.isValidElement(elements)).toBe(true)
+
+    const childElements = (
+      elements as React.ReactElement<{ children: React.ReactElement[] }>
+    ).props.children
+    const childKeys = childElements.map((el) => el?.key).filter(Boolean)
+
+    expect(cleanPrefix(childKeys[0] as string)).toBeLessThan(
+      cleanPrefix(childKeys[1] as string)
+    )
   })
 })